chore(deps): bump rusqlite, fs4, vite-plus and patch vulnerable transitives
Supersedes the Dependabot PRs that could not build as-is:

- rusqlite 0.34 -> 0.40: u64 lost its ToSql impl, so the Battle.net
  login-cache query now binds account_id_lo as i64 (SQLite stores
  integers as i64 anyway).
- fs4 0.13 -> 1.1: the fs_std module is gone and the sync trait moved
  to fs4::FileExt with lock/try_lock names mirroring std 1.89 file
  locking; try_lock now reports contention as TryLockError::WouldBlock
  instead of Ok(false).
- vite-plus 0.2.6 -> 0.2.8: bumped in devDependencies, the vite alias
  and the pnpm override together, with a fully re-resolved lockfile so
  the native rolldown bindings for all platforms land in the lockfile
  (the Dependabot bump missed them and vp install crashed on all OSes).

Security bumps for open Dependabot alerts:

- Cargo: serde_with 3.22, tar 0.4.46.
- root lockfile: picomatch 4.0.5.
- server lockfile: wrangler 4.122 pulling undici 7.29.0, ws 8.21.0,
  esbuild 0.28.1 and sharp 0.35.2; workers-types bumped to v5 to match
  wrangler's peer range.

+++ crates/accshift-core/src/platforms/battle_net.rs
@@ -208,7 +208,7 @@ fn current_battle_tag_from_cache() -> Result<Option<String>, String> {
         .map_err(|e| format!("Could not query Battle.net login cache: {e}"))?;
 
     let mut rows = statement
-        .query([account_id_lo])
+        .query([account_id_lo as i64])
         .map_err(|e| format!("Could not read Battle.net login cache: {e}"))?;
 
     let Some(row) = rows
